pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Nelson Carvalho <nelsonfilth@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2"
djangorestframework = "^3.15.2"
django-debug-toolbar = "^5.0.1"
django-extensions = "^3.2.3"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
faker = "^35.0.0"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# Correção para evitar erro de lockfile desatualizado
[tool.poetry.scripts]
preinstall = "bookstore.scripts:lock"
